+++ GP-Tool/include/plugin.h
@@ -17,10 +17,10 @@ struct Plugin
 ///////////////////////////////////////////////////////////
 // Some helper functions for JSON
 template <typename T>
-Json::Value jsonArray(T *vec, size_t N)
+Json::Value jsonArray(T *vec, uint64_t N)
 {
     Json::Value val(Json::arrayValue);
-    for (size_t k = 0; k < N; k++)
+    for (uint64_t k = 0; k < N; k++)
         val.append(vec[k]);
 
     return val;
