+++ Methods/include/align.h
@@ -6,7 +6,7 @@ namespace GPT
 {
     struct TransformData
     {
-        GP_API TransformData(uint32_t width, uint32_t height);
+        GP_API TransformData(uint64_t width, uint64_t height);
         GP_API TransformData(void) = default;
         GP_API ~TransformData(void) = default;
 
@@ -23,7 +23,7 @@ namespace GPT
     class Align
     {
     public:
-        GP_API Align(uint32_t nFrames, const MatXd *vim1, const MatXd *vim2);
+        GP_API Align(uint64_t nFrames, const MatXd *vim1, const MatXd *vim2);
         GP_API ~Align(void) = default;
 
         GP_API bool alignCameras(void);
@@ -47,7 +47,7 @@ namespace GPT
         Mat3d itrf;
         std::vector<double> global_energy;
 
-        void calcEnergy(const uint32_t id, const uint32_t nThr);
+        void calcEnergy(const uint64_t id, const uint64_t nThr);
         double weightTransRot(const VecXd &p);
         double weightScale(const VecXd &p);
